word-format-tool/requirements.txt
python-docx==0.8.11
streamlit==1.32.2
pandas==2.2.1